Fix test failing depending on run order
Location manager runs once with the fake parser before being reset and run
with the same fake parser again. In some cases this can cause the
locations in the fake parser to maintain a bad state. Isolating the first
initialization of the manager in a setup allows tests to run consistently
regardless of run order