Remove listeners able to prevent other listeners
It sat really uneasy with me, as I could see how that could cause hard to
notice bugs later on. Instead I made the UseEvent listeners inherit from a
base that checks the distance, so I don't have to do the check in every
listener. It means new actions/ non-use actions could work out of reach,
but that should be easier to spot than an event seemingly not firing
because it was blocked by someone else.